feat: update README.md with application details and add screenshot; refactor store imports to ShufflePickStore
